+++ src/resources/swagger.yaml
@@ -608,33 +608,6 @@ paths:
           description: Active submission is not in PENDING_APPROVAL state and cannot be reopned
         '500':
           $ref: '#/components/responses/ServerError'
-  /submission/program/{programId}/clinical/committed/tsv:
-    parameters:
-      - name: programId
-        description: the short name of the program
-        in: path
-        required: true
-        schema:
-          type: string
-    get:
-      summary: Get all commited clinical data for a program
-      description: returns all commited clinical and sample_registration data in submission ready tsv
-      responses:
-        '200':
-          description: 'A zip file'
-          content:
-            application/zip:
-              schema:
-                type: string
-                format: binary
-        '400':
-          description: Invalid programId provided
-        '401':
-          $ref: '#/components/responses/UnauthorizedError'
-        '403':
-          $ref: '#/components/responses/ForbiddenError'
-        '500':
-          $ref: '#/components/responses/ServerError'
 
   /submission/icgc-import/preprocess/{programId}:
     parameters:
@@ -820,6 +793,35 @@ paths:
       responses:
         '200':
           description: 'Donor with updated stats'
+  /clinical/program/{programId}/tsv-export:
+    parameters:
+      - name: programId
+        description: the short name of the program
+        in: path
+        required: true
+        schema:
+          type: string
+    get:
+      tags:
+        - Clinical Data
+      summary: Get all commited clinical data for a program
+      description: Returns a zip file with all commited clinical and sample_registration data in tsvs. Exported data may not be valid with current dictionary.
+      responses:
+        '200':
+          description: 'A zip file with tsv file(s) inside'
+          content:
+            application/zip:
+              schema:
+                type: string
+                format: binary
+        '400':
+          description: Invalid programId provided
+        '401':
+          $ref: '#/components/responses/UnauthorizedError'
+        '403':
+          $ref: '#/components/responses/ForbiddenError'
+        '500':
+          $ref: '#/components/responses/ServerError'
 
   /submission/schema:
     get:
