+++ src/submission/submission-api.ts
@@ -261,33 +261,6 @@ class SubmissionController {
     const donorsJson = await fsPromises.readFile(req.file.path, 'utf-8');
     return res.status(201).send(await submission.operations.adminAddDonors(JSON.parse(donorsJson)));
   }
-
-  @HasProgramWriteAccess((req: Request) => req.params.programId)
-  async downloadCommittedClinicalDataAsTsv(req: Request, res: Response) {
-    const programId = req.params.programId;
-    if (!programId) {
-      return ControllerUtils.badRequest(res, 'Invalid programId provided');
-    }
-
-    const data = await submission.operations.getAllCommittedClinicalData(programId);
-
-    res
-      .status(200)
-      .contentType('application/zip')
-      .attachment(`${programId}_all_clincial_data.zip`);
-
-    // prepare zip file
-    const zip = new AdmZip();
-    data.forEach(d => {
-      const tsvData = TsvUtils.parseObjectToTsv(d.records, d.schemaFields);
-      zip.addFile(
-        `${d.schemaEntityName}_v${d.schemaVersion}.tsv`,
-        Buffer.alloc(tsvData.length, tsvData),
-      );
-    });
-
-    res.send(zip.toBuffer());
-  }
 }
 
 const submissionSystemIsDisabled = async (res: Response) => {
