+++ src/submission/submission-service.ts
@@ -648,72 +648,6 @@ export namespace operations {
     );
   };
 
-  export async function getAllCommittedClinicalData(programId: string) {
-    if (!programId) throw new Error('Missing programId!');
-
-    // this query takes about 6sec for 20,000 donors with sample registration only running locally
-    const donors = await getDonors(programId); // get in batch??
-
-    // collect all records
-    const recordsMap: any = {};
-    donors.forEach(d => {
-      Object.values(ClinicalEntitySchemaNames).forEach(entity => {
-        let clincialInfoRecords;
-        if (entity === ClinicalEntitySchemaNames.REGISTRATION) {
-          clincialInfoRecords = generateSampleRegistrationRecordsFromDonor(d);
-        } else {
-          clincialInfoRecords = getClinicalEntitiesFromDonorBySchemaName(d, entity);
-        }
-        recordsMap[entity] = _.concat(recordsMap[entity] || [], clincialInfoRecords);
-      });
-    });
-
-    // map into object ready for api processing
-    const schemasWithFields = dictionaryManager.instance().getSchemasWithFields();
-    const schemaVersion = dictionaryManager.instance().getCurrent().version;
-    return Object.entries(recordsMap)
-      .map(([schemaEntityName, records]) => {
-        if (isEmpty(records)) return undefined;
-
-        const relevantSchemaWithFields = schemasWithFields.find(s => s.name === schemaEntityName);
-        if (!relevantSchemaWithFields) {
-          throw new Error(`Can't find schema ${schemaEntityName}, shouldn't be possbile`);
-        }
-
-        return {
-          schemaEntityName: relevantSchemaWithFields.name,
-          schemaVersion,
-          records,
-          schemaFields: relevantSchemaWithFields.fields,
-        };
-      })
-      .filter(notEmpty);
-  }
-
-  function generateSampleRegistrationRecordsFromDonor(
-    d: DeepReadonly<Donor>,
-  ): SubmittedRegistrationRecord[] {
-    const baseRegistrationRecord = {
-      program_id: d.programId,
-      submitter_donor_id: d.submitterId,
-      gender: d.gender,
-    };
-
-    return d.specimens
-      .map(sp =>
-        sp.samples.map(sm => ({
-          ...baseRegistrationRecord,
-          submitter_specimen_id: sp.submitterId,
-          specimen_tissue_source: sp.specimenTissueSource,
-          tumour_normal_designation: sp.tumourNormalDesignation,
-          specimen_type: sp.specimenType,
-          submitter_sample_id: sm.submitterId,
-          sample_type: sm.sampleType,
-        })),
-      )
-      .flat();
-  }
-
   /* *************** *
    * Private Methods
    * *************** */
